requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
aistudio-sdk==0.2.6
annotated-types==0.5.0
anyio==3.7.1
astor==0.8.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
Babel==2.14.0
bce-python-sdk==0.9.23
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
colorlog==6.8.2
cycler==0.11.0
datasets==2.13.2
decorator==5.1.1
dill==0.3.4
exceptiongroup==1.2.2
fastapi==0.103.2
filelock==3.12.2
Flask==2.2.5
flask-babel==3.1.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future==1.0.0
h11==0.14.0
h5py==3.8.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
jieba==0.42.1
Jinja2==3.1.4
joblib==1.3.2
kiwisolver==1.4.5
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mdurl==0.1.2
multidict==6.0.5
multiprocess==0.70.12.2
numpy==1.21.6
onnx==1.14.1
opt-einsum==3.3.0
packaging==24.0
paddle2onnx==1.1.0
paddlefsl==1.1.0
paddlenlp==2.7.2
paddlepaddle-gpu==2.5.2
pandas==1.3.5
Pillow==9.5.0
prettytable==3.7.0
protobuf==4.24.4
psutil==6.1.0
pyarrow==12.0.1
pybind11==2.13.6
pycryptodome==3.21.0
pydantic==2.5.3
pydantic_core==2.14.6
Pygments==2.17.2
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.1
rarfile==4.2
requests==2.31.0
rich==13.8.1
safetensors==0.4.5
scikit-learn==1.0.2
scipy==1.7.3
sentencepiece==0.2.0
seqeval==1.2.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.27.0
threadpoolctl==3.1.0
tool-helpers==0.1.2
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.7.1
urllib3==2.0.7
uvicorn==0.22.0
visualdl==2.5.3
wcwidth==0.2.13
Werkzeug==2.2.3
xxhash==3.5.0
yarl==1.9.4
zipp==3.15.0
